pyproject.toml
[project]
name = "sales-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.6",
    "duckduckgo-search>=8.1.1",
    "openai-agents>=0.17.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.3",
    "playwright>=1.59.0",
    "playwright-stealth>=2.0.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
]
